requirements.txt
aniso8601==7.0.0
astroid==2.2.5
attrs==19.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.229
botocore==1.12.229
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cymem==2.0.2
docutils==0.15.2
entrypoints==0.3
flake8==3.7.8
flake8-mypy==17.8.0
flake8-polyfill==1.0.2
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
geffnet==0.9.3
gensim==3.8.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.2
mypy==0.720
mypy-extensions==0.4.1
nltk==3.4.5
numpy==1.17.1
pandas==0.25.1
pep8-naming==0.8.2
Pillow==6.2.1
plac==0.9.6
preshed==2.0.1
psycopg2==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
smart-open==1.8.4
spacy==2.1.8
SQLAlchemy==1.3.8
srsly==0.1.0
thinc==7.0.8
https://download.pytorch.org/whl/cpu/torch-1.3.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
torchvision==0.4.2
tqdm==4.35.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.5
wrapt==1.11.2
